requirements.txt
django==4.2.10
django-crispy-forms==2.0
crispy-bootstrap5==2023.10
django-crispy-forms==2.0
djangorestframework==3.14.0